plotColouredExpression

the plotColouredExpression function plots an n-panel scatterplot of the gene pairs split by early, mid, and late in the sample ordering.

plotColouredExpression(
  scHOT,
  genes,
  genes_delimeter = "_",
  branches = NULL,
  ranked_by = NULL,
  subsetBranch = NULL,
  n = 3,
  fittedline = TRUE,
  assayName = NULL
)

Arguments

scHOT
A scHOT object.

genes
is either a single character string with a delimeter, or a length two character vector

genes_delimeter
is the delimeter to split into two gene names if genes is provided as a single character

branches
A character indicates that the colnames stored the branch information in colData

ranked_by
A character indicates that the colnames stored the ranking information of the cells in colData, such as trajectory time, if it is NULL, it will be ranked based on the branch information.

subsetBranch
subsetBranch is a character vector containing the names of the branches to be plotted. If NULL it will plot all branches

n
number of panels to split ranked samples into, default 3.

fittedline
logical default TRUE, add a lm straight line to the plot

assayName
the name of the assay that are used to plot.

Value

ggplot a ggplot object of scatterplots of expression split by sample ordering