scHOT

A wrap up function to perform scHOT

scHOT(
  scHOT,
  testingScaffold = NULL,
  weightMatrix = NULL,
  positionType = NULL,
  positionColData = NULL,
  nrow.out = NULL,
  averageAcrossTrajectoryTies = FALSE,
  higherOrderFunction = NULL,
  higherOrderFunctionType = NULL,
  numberPermutations = 1000,
  numberScaffold = 100,
  storePermutations = TRUE,
  higherOrderSummaryFunction = sd,
  parallel = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  usenperm = FALSE,
  nperm = 10000,
  maxDist = 0.1,
  plot = FALSE,
  verbose = TRUE,
  ...
)

Arguments

scHOT
A scHOT object

testingScaffold
A matrix with rows for each testing combination

weightMatrix
A matrix indicates the weight matrix for scHOT analysis

positionType
A string indicates the position type, either trajectory or spatial

positionColData
If positionType is "trajectory" then positionColData should be a sortable vector if positionType is "spatial" then positionColData should be a matrix type object. It should be stored in colData (SHILA to check).

nrow.out
SHILA to input

averageAcrossTrajectoryTies
SHILA to input

higherOrderFunction
A function object indicates the higher order function

higherOrderFunctionType
is weighted or unweighted, determines if there is a weighting argument in the higher order function

numberPermutations
The number of permutatuion, set as 1000 by default

numberScaffold
The number of scaffold, set as 100 by default

storePermutations
a logical flag on whether

higherOrderSummaryFunction
A functon indicates the higher order summary function (SHILA to check)

parallel
A logical input indicates whether run the permutation test using multiple cores in parallel.

BPPARAM
A BiocParallelParam class object from the BiocParallel package is used. Default is SerialParam().

usenperm
SHILA to write

nperm
number of permutation

maxDist
SHILA to write

plot
A logical input indicates whether the results are plotted

verbose
A logical input indicates whether the intermediate steps will be printed

...
parameters for function trajectoryWeightMatrix or spatialWeightMatrix